BackEnd/PythonProject1/requirements.txt
psycopg2-binary
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
deep-translator==1.11.4
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.1.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
langdetect==1.0.9
MarkupSafe==3.0.2
openai==1.93.0
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
pyjokes==0.8.3
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
wikipedia==1.4.0
